+++ erl/root_pid.go
@@ -1,5 +1,11 @@
 package erl
 
+import (
+	"fmt"
+
+	"github.com/uberbrodt/erl-go/erl/exitreason"
+)
+
 var rootPID PID
 
 // The idea here is that rootProc will always exist and just log
@@ -14,12 +20,18 @@ type rootProc struct{}
 
 // TODO: What if he dies for some reason? Does he need a supervisor?
 func (rp *rootProc) Receive(self PID, inbox <-chan any) error {
-	for {
-		msg := <-inbox
-
-		Logger.Printf("rootProc received: %+v", msg)
-
+	for anymsg := range inbox {
+		switch msg := anymsg.(type) {
+		case ExitMsg:
+			if !msg.Link {
+				Logger.Printf("RootPID received an exit signal with reason: %v", msg.Reason)
+				return exitreason.Exception(fmt.Errorf("RootPID received an exit signal with reason: %w", msg.Reason))
+			}
+		default:
+			Logger.Printf("rootProc received: %+v", msg)
+		}
 	}
+	return nil
 }
 
 func RootPID() PID {